Illustrate a reference's scope
A reference’s scope starts from where it is introduced and continues
through the last time that reference is used.

The scopes of the immutable references `r1` and `r2` end after the
`println!` where they are last used, which is before the mutable
reference `r3` is created. These scopes don’t overlap, so this code is
allowed.